+++ CHANGES.md
@@ -4,5 +4,6 @@ main
 ffmml
 -----
 
+- [UPDATE] Expose `Sample` type.
 - [UPDATE] Add `wav` feature
 - [FIX] Fix a bug that `MusicPlayer::take_last_error()` always returns `None`